requirements.txt
Flask~=3.0.3
#Flask-APScheduler~=1.13.1
Flask-Assets~=2.1.0
gunicorn==23.0.0
Werkzeug~=3.0.4
Flask-SocketIO~=5.4.1
requests~=2.32.3
#celery~=5.4.0
#SQLAlchemy~=2.0.36
#flask-sqlalchemy~=3.1.1
rich~=13.9.2
click~=8.1.7
python-dotenv~=1.0.1
paho-mqtt~=2.1.0
python-socketio~=5.11.4
websocket-client~=1.8.0
#rq~=2.0.0
redis~=5.2.1
WTForms~=3.2.1